Fix bump job (#128)
As part of the `bump` job, we are actually updating the `README.md` with
docgen. To do this, we need to install yarn. However, the orb's job
isn't actually installing that, so it's failing after creating the PR
but before pushing the changes to the VERSIONS.md and the README.md
files.

This copies the job from the orb to the repo in order to add an extra
step that installs yarn. We could potentially move this to the orb and
make it an optional parameter... But since this is only needed here for
now, I thought to leave it here. Lmk if you have other thoughts.