Fix `setAttributes` method to use parameters in the correct location (#263)
Currently the native processing assumes all attributes will be under a
`attributes` parameter in the method call, but that's actually not
allowed in the TS interface.

This PR changes the native processing to pick the attributes from the
root of the call data, passing the correct parameters to the native
side. This is not completely in line with our other methods in the SDK
API where we require a specific parameter key, but this solution allows
to avoid a breaking change.