Support capacitor V6 (#295)
This PR updates the SDK and sample app to Capacitor V6. Some people
might find this a breaking change depending on their android
configuration, since it bumps the android target/compile SDK versions
but this will be a requirement to post updates in the Google Play
console soon so it's a mandatory requirement.