Rename podspec to lowercase `RevenuecatPurchasesCapacitor.podspec` (#292)
Capacitor actually expects the podspec name to be lowercase so we can't
rename to `RevenueCatPurchasesCapacitor`. This seems to come from the
package name `@revenuecat/purchases-capacitor`. This moves everywhere we
use this as uppercase to lowercase.

I've tested both a clean installation and an update after this change in
our local purchase tester and seems to work fine