Capacitor V7 Support (#387)
This adds support for Capacitor V7 which includes some breaking changes
like:
- Increased android min and target sdk versions. 
- Increased iOS deployment target to iOS 14.
- Bunch of dependency updates

This PR is the result of running the provided migrator executables in
the SDK and our sample app. Most things worked out of the box with our
sample app.

To see more details please check the official documentation:
https://capacitorjs.com/docs/updating/7-0